Config in process

+++ wireframe/src/main/java/ru/fit/g14201/chirikhin/wireframe/bspline/Point.java
@@ -1,19 +1,19 @@
 package ru.fit.g14201.chirikhin.wireframe.bspline;
 
 public class Point {
-    private final double x;
-    private final double y;
+    private final float x;
+    private final float y;
 
-    public Point(double x, double y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 }
